combine tests for easy maintenance
