+++ tests/pytorch/test_transformerengine.py
@@ -5,6 +5,7 @@
 import torch
 import pytest
 
+from transformer_engine.pytorch.fp8 import fp8_autocast
 from transformer_engine.pytorch.utils import (
     init_method_normal,
     scaled_init_method_normal,
@@ -15,6 +16,29 @@ from transformer_engine.pytorch import (
     LayerNormMLP,
     TransformerLayer,
 )
+from transformer_engine.common import recipe
+
+# Only run FP8 tests on H100.
+fp8_available = torch.cuda.get_device_properties(torch.cuda.current_device()).major >= 9
+
+
+def custom_amax_to_scale(
+    amax: torch.Tensor,
+    scale: torch.Tensor,
+    fp8_max: torch.Tensor,
+    recipe: recipe.DelayedScaling,
+) -> torch.Tensor:
+    """Custom func to test recipe."""
+    sf = fp8_max / amax
+    sf = torch.where(amax > 0.0, sf, scale)
+    sf = torch.where(torch.isfinite(amax), sf, scale)
+
+    return sf
+
+
+def custom_amax_compute(amax_history: torch.Tensor) -> torch.Tensor:
+    """Custom func to test recipe."""
+    return torch.min(amax_history, dim=0).values
 
 
 class ModelConfig:
@@ -33,6 +57,35 @@ model_configs = {
     "126m": ModelConfig(768, 1e-5, 12, 64, 12, 2048),
 }
 
+fp8_recipes = [
+    None, # Handles non-FP8 case
+    recipe.DelayedScaling(0, 1, recipe.Format.E4M3),
+    recipe.DelayedScaling(0, 1, recipe.Format.HYBRID),
+    recipe.DelayedScaling(
+        0, 1, recipe.Format.E4M3, override_linear_precision=(False, False, True)
+    ),
+    recipe.DelayedScaling(
+        0, 1, recipe.Format.E4M3, amax_history_len=16, amax_compute_algo="most_recent"
+    ),
+    recipe.DelayedScaling(
+        0, 1, recipe.Format.E4M3, amax_history_len=16, amax_compute_algo="max"
+    ),
+    recipe.DelayedScaling(
+        0,
+        1,
+        recipe.Format.E4M3,
+        amax_history_len=16,
+        amax_compute_algo=custom_amax_compute,
+    ),
+    recipe.DelayedScaling(
+        0,
+        1,
+        recipe.Format.E4M3,
+        amax_history_len=16,
+        scaling_factor_compute_algo=custom_amax_to_scale,
+    ),
+]
+
 param_types = [torch.float32, torch.bfloat16, torch.float16]
 
 batch_sizes = [1, 2]
@@ -40,20 +93,15 @@ batch_sizes = [1, 2]
 all_boolean = [True, False]
 
 
-
 def _disable_wgrads(block):
     for p in block.parameters():
         p.requires_grad = False
 
 
-def _test_sanity_e2e_amp(block, bs, dtype, config, skip_wgrad):
-    if dtype == torch.bfloat16 and not torch.cuda.is_bf16_supported():
-        return
-
+def _test_sanity_e2e_amp(block, bs, dtype, config, fp8_recipe, skip_wgrad):
     te_inp_hidden_states = torch.randn(
         config.seq_len, bs, config.hidden_size, dtype=torch.float32, requires_grad=True
     ).cuda()
-
     te_inp_attn_mask = (
         torch.rand(
             (
@@ -70,8 +118,10 @@ def _test_sanity_e2e_amp(block, bs, dtype, config, skip_wgrad):
     if skip_wgrad:
         _disable_wgrads(block)
 
+    use_fp8 = fp8_recipe is not None
     with torch.cuda.amp.autocast(enabled=True, dtype=dtype):
-        te_out = block(te_inp_hidden_states, te_inp_attn_mask)
+        with fp8_autocast(enabled=use_fp8, fp8_recipe=fp8_recipe):
+            te_out = block(te_inp_hidden_states, te_inp_attn_mask)
         loss = te_out.sum()
 
     assert te_out.dtype == dtype
@@ -79,7 +129,7 @@ def _test_sanity_e2e_amp(block, bs, dtype, config, skip_wgrad):
     torch.cuda.synchronize()
 
 
-def _test_sanity_e2e(block, bs, dtype, config, skip_wgrad):
+def _test_sanity_e2e(block, bs, dtype, config, fp8_recipe, skip_wgrad):
     te_inp_hidden_states = torch.randn(
         config.seq_len, bs, config.hidden_size, dtype=dtype, requires_grad=True
     ).cuda()
@@ -99,13 +149,15 @@ def _test_sanity_e2e(block, bs, dtype, config, skip_wgrad):
     if skip_wgrad:
         _disable_wgrads(block)
 
-    te_out = block(te_inp_hidden_states, te_inp_attn_mask)
+    use_fp8 = fp8_recipe is not None
+    with fp8_autocast(enabled=use_fp8, fp8_recipe=fp8_recipe):
+        te_out = block(te_inp_hidden_states, te_inp_attn_mask)
     loss = te_out.sum()
     loss.backward()
     torch.cuda.synchronize()
 
 
-def _test_sanity_e2e_T5(block, bs, dtype, config, skip_wgrad):
+def _test_sanity_e2e_T5(block, bs, dtype, config, fp8_recipe, skip_wgrad):
     te_inp_hidden_states = torch.randn(
         config.seq_len, bs, config.hidden_size, dtype=dtype, requires_grad=True
     ).cuda()
@@ -125,15 +177,17 @@ def _test_sanity_e2e_T5(block, bs, dtype, config, skip_wgrad):
     if skip_wgrad:
         _disable_wgrads(block)
 
-    te_out = block(
-        te_inp_hidden_states, te_inp_attn_mask, encoder_output=te_inp_hidden_states
-    )
+    use_fp8 = fp8_recipe is not None
+    with fp8_autocast(enabled=use_fp8, fp8_recipe=fp8_recipe):
+        te_out = block(
+            te_inp_hidden_states, te_inp_attn_mask, encoder_output=te_inp_hidden_states
+        )
     loss = te_out.sum()
     loss.backward()
     torch.cuda.synchronize()
 
 
-def _test_sanity_common(block, bs, dtype, config, skip_wgrad):
+def _test_sanity_common(block, bs, dtype, config, fp8_recipe, skip_wgrad):
     te_inp = torch.randn(
         config.seq_len, bs, config.hidden_size, dtype=dtype, requires_grad=True
     ).cuda()
@@ -141,7 +195,9 @@ def _test_sanity_common(block, bs, dtype, config, skip_wgrad):
     if skip_wgrad:
         _disable_wgrads(block)
 
-    te_out = block(te_inp)
+    use_fp8 = fp8_recipe is not None
+    with fp8_autocast(enabled=use_fp8, fp8_recipe=fp8_recipe):
+        te_out = block(te_inp)
     if isinstance(te_out, tuple):
         te_out = te_out[0]
     loss = te_out.sum()
@@ -151,10 +207,14 @@ def _test_sanity_common(block, bs, dtype, config, skip_wgrad):
 
 @pytest.mark.parametrize("dtype", param_types)
 @pytest.mark.parametrize("bs", batch_sizes)
+@pytest.mark.parametrize("fp8_recipe", fp8_recipes)
 @pytest.mark.parametrize("model", model_configs.keys())
 @pytest.mark.parametrize("skip_wgrad", all_boolean)
 @pytest.mark.parametrize("zero_centered_gamma", all_boolean)
-def test_sanity_layernorm_linear(dtype, bs, model, skip_wgrad, zero_centered_gamma):
+def test_sanity_layernorm_linear(dtype, bs, fp8_recipe, model, skip_wgrad, zero_centered_gamma):
+    if fp8_recipe is not None and not fp8_available:
+        pytest.skip("FP8 device not available.")
+
     config = model_configs[model]
 
     sigma = 0.023
@@ -171,14 +231,18 @@ def test_sanity_layernorm_linear(dtype, bs, model, skip_wgrad, zero_centered_gam
         .to(dtype=dtype)
         .cuda()
     )
-    _test_sanity_common(block, bs, dtype, config, skip_wgrad)
+    _test_sanity_common(block, bs, dtype, config, fp8_recipe, skip_wgrad)
 
 
 @pytest.mark.parametrize("dtype", param_types)
 @pytest.mark.parametrize("bs", batch_sizes)
+@pytest.mark.parametrize("fp8_recipe", fp8_recipes)
 @pytest.mark.parametrize("model", model_configs.keys())
 @pytest.mark.parametrize("skip_wgrad", all_boolean)
-def test_sanity_linear(dtype, bs, model, skip_wgrad):
+def test_sanity_linear(dtype, bs, fp8_recipe, model, skip_wgrad):
+    if fp8_recipe is not None and not fp8_available:
+        pytest.skip("FP8 device not available.")
+
     config = model_configs[model]
 
     sigma = 0.023
@@ -191,16 +255,19 @@ def test_sanity_linear(dtype, bs, model, skip_wgrad):
         .to(dtype=dtype)
         .cuda()
     )
-
-    _test_sanity_common(block, bs, dtype, config, skip_wgrad)
+    _test_sanity_common(block, bs, dtype, config, fp8_recipe, skip_wgrad)
 
 
 @pytest.mark.parametrize("dtype", param_types)
 @pytest.mark.parametrize("bs", batch_sizes)
+@pytest.mark.parametrize("fp8_recipe", fp8_recipes)
 @pytest.mark.parametrize("model", model_configs.keys())
 @pytest.mark.parametrize("skip_wgrad", all_boolean)
 @pytest.mark.parametrize("zero_centered_gamma", all_boolean)
-def test_sanity_layernorm_mlp(dtype, bs, model, skip_wgrad, zero_centered_gamma):
+def test_sanity_layernorm_mlp(dtype, bs, fp8_recipe, model, skip_wgrad, zero_centered_gamma):
+    if fp8_recipe is not None and not fp8_available:
+        pytest.skip("FP8 device not available.")
+
     config = model_configs[model]
 
     sigma = 0.023
@@ -219,15 +286,19 @@ def test_sanity_layernorm_mlp(dtype, bs, model, skip_wgrad, zero_centered_gamma)
         .to(dtype=dtype)
         .cuda()
     )
-    _test_sanity_common(block, bs, dtype, config, skip_wgrad)
+    _test_sanity_common(block, bs, dtype, config, fp8_recipe, skip_wgrad)
 
 
 @pytest.mark.parametrize("dtype", param_types)
 @pytest.mark.parametrize("bs", batch_sizes)
+@pytest.mark.parametrize("fp8_recipe", fp8_recipes)
 @pytest.mark.parametrize("model", model_configs.keys())
 @pytest.mark.parametrize("skip_wgrad", all_boolean)
 @pytest.mark.parametrize("zero_centered_gamma", all_boolean)
-def test_sanity_gpt(dtype, bs, model, skip_wgrad, zero_centered_gamma):
+def test_sanity_gpt(dtype, bs, fp8_recipe, model, skip_wgrad, zero_centered_gamma):
+    if fp8_recipe is not None and not fp8_available:
+        pytest.skip("FP8 device not available.")
+
     config = model_configs[model]
 
     sigma = 0.023
@@ -253,15 +324,19 @@ def test_sanity_gpt(dtype, bs, model, skip_wgrad, zero_centered_gamma):
         .cuda()
     )
 
-    _test_sanity_e2e(block, bs, dtype, config, skip_wgrad)
+    _test_sanity_e2e(block, bs, dtype, config, fp8_recipe, skip_wgrad)
 
 
 @pytest.mark.parametrize("dtype", param_types)
 @pytest.mark.parametrize("bs", batch_sizes)
+@pytest.mark.parametrize("fp8_recipe", fp8_recipes)
 @pytest.mark.parametrize("model", model_configs.keys())
 @pytest.mark.parametrize("skip_wgrad", all_boolean)
 @pytest.mark.parametrize("zero_centered_gamma", all_boolean)
-def test_sanity_bert(dtype, bs, model, skip_wgrad, zero_centered_gamma):
+def test_sanity_bert(dtype, bs, fp8_recipe, model, skip_wgrad, zero_centered_gamma):
+    if fp8_recipe is not None and not fp8_available:
+        pytest.skip("FP8 device not available.")
+
     config = model_configs[model]
 
     sigma = 0.023
@@ -287,15 +362,19 @@ def test_sanity_bert(dtype, bs, model, skip_wgrad, zero_centered_gamma):
         .cuda()
     )
 
-    _test_sanity_e2e(block, bs, dtype, config, skip_wgrad)
+    _test_sanity_e2e(block, bs, dtype, config, fp8_recipe, skip_wgrad)
 
 
 @pytest.mark.parametrize("dtype", param_types)
 @pytest.mark.parametrize("bs", batch_sizes)
+@pytest.mark.parametrize("fp8_recipe", fp8_recipes)
 @pytest.mark.parametrize("model", model_configs.keys())
 @pytest.mark.parametrize("skip_wgrad", all_boolean)
 @pytest.mark.parametrize("zero_centered_gamma", all_boolean)
-def test_sanity_T5(dtype, bs, model, skip_wgrad, zero_centered_gamma):
+def test_sanity_T5(dtype, bs, fp8_recipe, model, skip_wgrad, zero_centered_gamma):
+    if fp8_recipe is not None and not fp8_available:
+        pytest.skip("FP8 device not available.")
+
     config = model_configs[model]
 
     sigma = 0.023
@@ -322,14 +401,18 @@ def test_sanity_T5(dtype, bs, model, skip_wgrad, zero_centered_gamma):
         .cuda()
     )
 
-    _test_sanity_e2e_T5(block, bs, dtype, config, skip_wgrad)
+    _test_sanity_e2e_T5(block, bs, dtype, config, fp8_recipe, skip_wgrad)
 
 
 @pytest.mark.parametrize("dtype", param_types)
 @pytest.mark.parametrize("bs", batch_sizes)
+@pytest.mark.parametrize("fp8_recipe", fp8_recipes)
 @pytest.mark.parametrize("model", model_configs.keys())
 @pytest.mark.parametrize("skip_wgrad", all_boolean)
-def test_sanity_amp_and_nvfuser(dtype, bs, model, skip_wgrad):
+def test_sanity_amp_and_nvfuser(dtype, bs, fp8_recipe, model, skip_wgrad):
+    if fp8_recipe is not None and not fp8_available:
+        pytest.skip("FP8 device not available.")
+
     config = model_configs[model]
 
     sigma = 0.023
@@ -352,14 +435,18 @@ def test_sanity_amp_and_nvfuser(dtype, bs, model, skip_wgrad):
         .cuda()
     )
 
-    _test_sanity_e2e_amp(block, bs, dtype, config, skip_wgrad)
+    _test_sanity_e2e_amp(block, bs, dtype, config, fp8_recipe, skip_wgrad)
 
 
 @pytest.mark.parametrize("dtype", param_types)
 @pytest.mark.parametrize("bs", batch_sizes)
+@pytest.mark.parametrize("fp8_recipe", fp8_recipes)
 @pytest.mark.parametrize("model", model_configs.keys())
 @pytest.mark.parametrize("skip_wgrad", all_boolean)
-def test_sanity_drop_path(dtype, bs, model, skip_wgrad):
+def test_sanity_drop_path(dtype, bs, fp8_recipe, model, skip_wgrad):
+    if fp8_recipe is not None and not fp8_available:
+        pytest.skip("FP8 device not available.")
+
     config = model_configs[model]
 
     sigma = 0.023
@@ -385,14 +472,18 @@ def test_sanity_drop_path(dtype, bs, model, skip_wgrad):
         .cuda()
     )
 
-    _test_sanity_e2e(block, bs, dtype, config, skip_wgrad)
+    _test_sanity_e2e(block, bs, dtype, config, fp8_recipe, skip_wgrad)
 
 
 @pytest.mark.parametrize("dtype", param_types)
 @pytest.mark.parametrize("bs", batch_sizes)
+@pytest.mark.parametrize("fp8_recipe", fp8_recipes)
 @pytest.mark.parametrize("model", model_configs.keys())
 @pytest.mark.parametrize("skip_wgrad", all_boolean)
-def test_sanity_fused_qkv_params(dtype, bs, model, skip_wgrad):
+def test_sanity_fused_qkv_params(dtype, bs, fp8_recipe, model, skip_wgrad):
+    if fp8_recipe is not None and not fp8_available:
+        pytest.skip("FP8 device not available.")
+
     config = model_configs[model]
 
     sigma = 0.023
@@ -418,4 +509,4 @@ def test_sanity_fused_qkv_params(dtype, bs, model, skip_wgrad):
         .cuda()
     )
 
-    _test_sanity_e2e(block, bs, dtype, config, skip_wgrad)
+    _test_sanity_e2e(block, bs, dtype, config, fp8_recipe, skip_wgrad)
